Fix build errors on non-CoreCLR runtimes
